--- a/knex/migrations/20180522000003_titres_etapes_substances_emprises.js
+++ b/knex/migrations/20180522000003_titres_etapes_substances_emprises.js
@@ -1,12 +1,16 @@
 exports.up = knex => {
   return knex.schema.createTable('titresSubstances', table => {
-    table.string('titreEtapeId', 128).notNullable()
+    table.string('titreEtapeId', 128).notNullable().index()
     table
       .foreign('titreEtapeId')
       .references('titresEtapes.id')
       .onUpdate('CASCADE')
       .onDelete('CASCADE')
-    table.string('substanceId', 4).references('substances.id').notNullable()
+    table
+      .string('substanceId', 4)
+      .index()
+      .references('substances.id')
+      .notNullable()
     table.boolean('connexe')
     table.integer('ordre')
     table.primary(['titreEtapeId', 'substanceId'])