--- a/knex/migrations/20180522000004_titres_etapes_utilisateurs_titulaires_amodiataires_administrations.js
+++ b/knex/migrations/20180522000004_titres_etapes_utilisateurs_titulaires_amodiataires_administrations.js
@@ -1,7 +1,7 @@
 exports.up = knex => {
   return knex.schema
     .createTable('titresTitulaires', table => {
-      table.string('titreEtapeId', 128).notNullable()
+      table.string('titreEtapeId', 128).notNullable().index()
       table
         .foreign('titreEtapeId')
         .references('titresEtapes.id')
@@ -9,6 +9,7 @@ exports.up = knex => {
         .onDelete('CASCADE')
       table
         .string('entrepriseId', 64)
+        .index()
         .references('entreprises.id')
         .onUpdate('CASCADE')
         .onDelete('CASCADE')
@@ -19,11 +20,13 @@ exports.up = knex => {
     .createTable('titresAmodiataires', table => {
       table
         .string('titreEtapeId', 128)
+        .index()
         .references('titresEtapes.id')
         .notNullable()
         .onDelete('CASCADE')
       table
         .string('entrepriseId', 64)
+        .index()
         .references('entreprises.id')
         .onUpdate('CASCADE')
         .onDelete('CASCADE')
@@ -33,7 +36,7 @@ exports.up = knex => {
     })
 
     .createTable('titresAdministrationsGestionnaires', table => {
-      table.string('titreId', 128).notNullable()
+      table.string('titreId', 128).notNullable().index()
       table
         .foreign('titreId')
         .references('titres.id')
@@ -41,6 +44,7 @@ exports.up = knex => {
         .onDelete('CASCADE')
       table
         .string('administrationId', 64)
+        .index()
         .references('administrations.id')
         .notNullable()
         .onUpdate('CASCADE')
@@ -49,7 +53,7 @@ exports.up = knex => {
       table.primary(['titreId', 'administrationId'])
     })
     .createTable('titresAdministrationsLocales', table => {
-      table.string('titreEtapeId', 128).notNullable()
+      table.string('titreEtapeId', 128).notNullable().index()
       table
         .foreign('titreEtapeId')
         .references('titresEtapes.id')
@@ -57,6 +61,7 @@ exports.up = knex => {
         .onDelete('CASCADE')
       table
         .string('administrationId', 64)
+        .index()
         .references('administrations.id')
         .notNullable()
         .onUpdate('CASCADE')