--- a/knex/migrations/20180522000006_titres_etapes_communes.js
+++ b/knex/migrations/20180522000006_titres_etapes_communes.js
@@ -1,12 +1,12 @@
 exports.up = knex =>
   knex.schema.createTable('titresCommunes', table => {
-    table.string('titreEtapeId', 128).notNullable()
+    table.string('titreEtapeId', 128).notNullable().index()
     table
       .foreign('titreEtapeId')
       .references('titresEtapes.id')
       .onUpdate('CASCADE')
       .onDelete('CASCADE')
-    table.string('communeId', 8).notNullable().references('communes.id')
+    table.string('communeId', 8).notNullable().index().references('communes.id')
     table.integer('surface')
     table.primary(['titreEtapeId', 'communeId'])
   })