--- a/knex/migrations/20180522000009_titres_references.js
+++ b/knex/migrations/20180522000009_titres_references.js
@@ -1,12 +1,16 @@
 exports.up = knex => {
   return knex.schema.createTable('titresReferences', table => {
-    table.string('titreId', 128).notNullable()
+    table.string('titreId', 128).notNullable().index()
     table
       .foreign('titreId')
       .references('titres.id')
       .onUpdate('CASCADE')
       .onDelete('CASCADE')
-    table.string('typeId', 3).references('referencesTypes.id').notNullable()
+    table
+      .string('typeId', 3)
+      .index()
+      .references('referencesTypes.id')
+      .notNullable()
     table.string('nom')
     table.primary(['titreId', 'typeId', 'nom'])
   })