Suite à l'étude des temps d'exécution des requêtes SQL, nous avons
remarqué qu'une requête était particulièrement lente sur la table
titres_points et la jointure sur titre_etape_id.

Avec la création de l'index, le temps de traitement passe de 220 ms
à 23 ms sur la base de prod.

La majorité des tables contiennent moins de 7k enregistrements sauf :
- titres_points_references: 19k
- titres_points: 18k
- titres_communes: 14k
- titres_etapes: 11k

Nous pourrions uniquement créer des index pour ces tables toutefois,
il nous semble pertinent de systématiquement utiliser des
index sur le FK dans la mesure où :
- l'espace de stockage n'est pas une contrainte
- le nombre d'ajout/suppression est faible.

Des frameworks tels que Django utilise ce comportement par défaut.

L'impact sur la réponse du serveur à la requête de la page d'accueil est
d'environ 270 ms (950 ms à 680 ms).

PS : la création de l'index GIST est dans la PR associée au périmètre.