@risseraka 2020-06-24
![](https://github.trello.services/images/mini-trello-icon.png) [perf: index de base de données](https://trello.com/c/mgTun1uK/369-perf-index-de-base-de-donn%C3%A9es)